requirements.txt
tenacity==8.2.3
python-dotenv==1.0.0
requests==2.31.0
yfinance==0.2.36


fredapi==0.5.2  # Latest available version
alpha-vantage==2.3.1

pandera==0.16.0
pandas==2.1.1
arch==6.2.0
scipy==1.10.1
causalnex==0.12.1
statsmodels==0.14.0
numpy==1.24.4

python-dotenv==1.0.0
aiohttp==3.8.6  # Explicitly add this
